put 'd-i' (di-kernel, di-initrd) into same stream as ephemeral
in talking to Oleg today, this makes his implementation of the reader
side easier.  And also, maas doesn't have to worry about the possibilty
that there is no di-kernel for a given version.

Instead we're forced to deal with that on the server side.
This commit includes a hack to re-use the di-kernel from one release
to the next.  That will force clients using this as a source to
handle the fact that a single path might be used across multiple
versions.

bzr-revno: 14