become aware of 'image-format', configure suffix
This makes the "image-format" implemenation a bit nicer.
We now have the data in netboot data for 'image-format'.
The only 'image-format' at the moment is 'xgene'.

This also extends the config to to accomodate.  Instead of the kernel arrays
being:
  krel, karch, primary-subarch, flavor, kernel-package, [subarches]
They are now:
  krel, karch, primary-subarch, flavor, kernel-package, [subarches], {data}

The 'data' is a dict of key/value pairs that mean something to
different parts of this code.

 suffix: a suffix for the items generated
 di-image-flavor: the image-flavor to pull from d-i
 kihelper: a program to run to "fix" initrd and kernel that
           come out of kpack-from-image

Previously, we did not consider the need for multiple different
"image formats" of the same
   (kernel-release, kernel-arch, primary-subarch, kernel-flavor)

Essentially, those tokens were our key for a unique kernel/initrd as it
was designed.  This basically adds 'image-format' to the list there,
defaulting to 'default'.

bzr-revno: 133.1.7