add support for linux-keystone (subarch=slayton)
The one tag-along here is resizing the filesystem in the mounted
image to accomodate the fact that in that image,
apt-get install of linux-keystone ended up doing::
  After this operation, 205 MB of additional disk space will be used.

And it would not fit.  That, coupled with the temporary initramfs
space caused out of space. Another thing we could do would be to bind mount
something over /tmp.

bzr-revno: 169