fix importing of armhf hwe-u kernels
Here we have to know that armhf kernels are published in
with paths like 'utopic-generic' or 'utopic-generic-lpae' where
on other releases, we would have seen 'utopic-netboot'.
Likely this is simply because we're seeing hwe kernels for non-intel
arches and also for additional flavors.

Without this change, the armhf kernels were being seen as trusty
kernels and would cause a name collision.

bzr-revno: 175