separate distro_info logic out.
Currenlty we'd have an issue if you just ran meph2-cloudimg-sync
without a release qualifier. The reason was that 'saucy' is in
conf/meph-v2.yaml but not in RELEASES.

now, RELEASES gets all releases.
at some point we should determine what to do about conf/meph-v2.yaml
having unsupported releases in it.

bzr-revno: 201