cloudimg-sync: support 'enable_proposed' for mining d-i data from -proposed
This does not yet support adding -proposed to an image's apt sources
so that you could find kernel images from -proposed.

It does, though support getting d-i data fom -proposed mirror locations.
This is sufficient for the case (that is true right now) where
linux-lts-vivid is available in -updates of trusty, but the d-i kernels are
still in -proposed.

To enable this, just add 'enable_proposed: True' in conf/meph-v2.yaml.

bzr-revno: 221