tox: run make 'meph2-util' in minimal python3 environment
The last merge caused meph2-util to fail to run on our build environment.
This adds a very minimal test that would have actually failed there.

We just run meph2-util in the tox env.  The imports of yaml and of
requests would have caused this to fail previously.

bzr-revno: 237