support input as '-root.tar.gz' (or .tar.xz)
In order to support this, we basically take the tar
and populate an image.  This was easier todo as kpack-from-image
takes an image also.

The other change/benefit of doing it this way is that we're now
copying the content of the filesystem into a fresh disk image
so it will have less (or no) dirty files as all that is done
is a rsync into it.

The one missing need right now is the 'padding' that was done for
ppc64el.  We dont have a way yet to make the image "minimum + 100M".

bzr-revno: 257.1.12