add '--rebuild' flag to cloudimg-sync
This allows you to build YYYYMMDD.1 if there is already
a maas image of YYYYMMDD for the latest cloud image YYYYMMDD.

Essentially, allowing a forced rebuild of a maas image even
if the sync doesn't think there is one needed.  I added this
as the 'maas-images' codebase created a bug in the 20160119
maas image, and wanted to create a 20160119.1 with the fix there.

that can now be accomplished with:
   meph2-cloudimg-sync -vvv --rebuild 20160119:20160119.1 \
     /tmp/out.d arch=amd64 release=trusty

bzr-revno: 263