make 'version' data in conf/meph-v2.yaml a string
yaml interprets 'version: 15.10' as a float (reasonably so).
the result was that new image builds were getting '15.1' as a version.
and even that propogating to the product name:
   com.ubuntu.maas.daily:v2:boot:15.1:amd64:hwe-w
rather than
   com.ubuntu.maas.daily:v2:boot:15.10:amd64:hwe-w

The fix here changes the config, and then also complains if it is a float
to cause early failure if this happened again later.

bzr-revno: 264