commit working tests
 * support s390x
 * call dtb files initrd_flavor 'netboot', altough they're not specific
   just for simplicitly sake. we're really only supporting that now
   not using 'gtk' or 'xen'
 * fix expected results for many hwe kernels

the get_file_item_data should be much easier to read now.  it strives to be
less magical. we do have more hard coded names (known image formats and kernel
flavors).  But that does make it much easier for someone to just add a
'kernel.ubuntu' and have it dtrt.

bzr-revno: 272.1.5