kpack-from-image: add drivers/net/usb for usb network devices
This adds ~ 1.5M more of drivers for usb network devices to
the initramfs.  The use case we're most interested in usb NICs for
is the "orange box" where each system only has 1 nic, but some 
workloads want 2.  We want maas to see the nic in commissioning 
and in deployment.

LP: #1552378
bzr-revno: 286