revert 'Add clean-unsupported command to meph2-utils'
the revision 301 added a dependency for meph2-utils on 
python3-distro-info which is not available on one of the build systems.
I'm reverting that commit and also removing 'distro-info' module
from the tox test environment for jerff so that if simply re-added
the tox test would correctly fail.

I've opened bug 1589559 to address re-adding the content in a
way that does not break things.

bzr-revno: 302