no longer symlink /etc/network/interfaces to /run in yakkety
bug 1534205 is fixed in xenial, but since it was done so close to GA
we did not change the images.  We could do that at a later point in time
with some focused testing.  For yakkety, though the symlink from
 /etc/network/interfaces -> ../../run/network/dynamic-interfaces
is not necessary.

This is the last change in image to the the maas image build process.

bzr-revno: 303