Disable metadata_csum when creating ext4 filesystems
In yakkety, we have sync'd an e2fsprogs version that supports metadata
checksumming in ext4 and enables the feature by default.  Unfortunately, some
Launchpad buildds (specifically those for armhf and one for powerpc) are
running on a kernel that is too old to support this, which means an unusable
ext4 partition is created.  To work around this issue, we disable the
metadata_csum option when we create an ext4 filesystem.

The Launchpad team are working to upgrade the offending build hosts, so we
should be able to back this change out before yakkety releases.

bzr-revno: 304.1.1