Import and publish bootloaders from a yaml file.
This allows the import subcommand to accept a yaml file containing a list
of bootloaders. The latest version of the bootloader is downloaded from
the Ubuntu archive. Files can be copied out of the bootloader or they can
be processed with grub-mkimage.

This also fixes a bug with the import command where if any version of a product
existed no newer version would be imported. This will allow more up-to-date
CentOS images to be published.

bzr-revno: 304.2.1