netinst.py: handle identical data published at different paths
the process that builds the d-i netboot data began publishing a
kernel and initramfs at different paths in 20101020ubuntu318.39.
For example, under
   /ubuntu-ports/dists/trusty-updates/main/installer-arm64/\
      20101020ubuntu318.40/images/xenial-generic/netboot/

There are now the same kernel and initrd and kernel at
  xenial-generic/netboot/initrd.gz
  xenial-generic/netboot/ubuntu-installer/arm64/initrd.gz
  xenial-generic/netboot/ubuntu-installer/arm64/linux
  xenial-generic/netboot/vmlinuz

That would cause a Name Collision in mine_md error on these files.

The change to address this is to raise exception only if
a name collision has different data than the existing.  Also
we specifically handle 'dtb' as currently the basename (derived
from the url) is used there.

bzr-revno: 310.1.1