move the 'import' subcommand out of meph2-util
Import did not fit as 'util' all that well anyway.  This moves 'import'
to its own standalone command, and exposes 'meph2-import' as a command
to call it.  It also moves some other code around so that it is easily
callable (moving many things to meph2.util).

It is still available with a deprecation warning as 'meph2-util import'.

The problem this fixes is in the environment that 'meph2-util' needs
to run in.  It is a python3.2 environment, and there is no 'lzma' or
backports.lzma available there.

When 'meph2-util import' was added, the build environment for some
of the uses of meph2-util was modified to use pip to install backports.
However, some of those jenkins jobs were *not*, resulting in failure
of some jobs such as the cleaner (LP: #1627739).

bzr-revno: 332.1.1