Add hwe-16.04-edge and hwe-16.04-lowlatency-edge kernels to V3 stream.
This adds the hwe-16.04-edge and hwe-16.04-lowlatency-edge kernels to the V3
stream. As the new names caused the yaml columns to be unaligned this adds
extra spaces to align all columns making the meph-v3.yaml file easier to read
and edit. While testing I discovered a small bug where edge was being added
before the kflavor in the product name as well as a lint error.

bzr-revno: 351.1.1