netinst: support d-i hwe kernels from rolling hwe
For LTS releases prior to xenial, the hardware enablement kernel was
specifically tied to a subsequent release.  Thus, for trusty we had
d-i paths like:

 device-tree/apm-merlin.dtb
 wily-device-tree/apm-merlin.dtb
 xenial-device-tree/apm-merlin.dtb

 netboot/ubuntu-installer/i386/linux
 wily-netboot/ubuntu-installer/i386/linux
 xenial-netboot/ubuntu-installer/i386/linux

With xenial, the hwe kernel is not specifically tied to a release, but
rather "rolls".  So we have new paths that look like this:

 hwe-device-tree/apm-merlin.dtb
 hwe-netboot/ubuntu-installer/i386/linux

For those the 'kernel release' data is set to simply "hwe".

The unit test data tests/unittests/file_item_data.json is updated here
and sorted with the following:
  $ cat sort-json.py
  import json, sys
  with open(sys.argv[1], "rb") as fp:
      data = json.loads(fp.read().encode())
  with open(sys.argv[1], "wb") as fp:
      fp.write(json.dumps(data, indent=1, sort_keys=True,
                        separators=(",", ": ")).encode("utf-8") + "\n")
  $ sort-json.py tests/unittests/file_item_data.json


LP: #1665376
bzr-revno: 354.1.1