When using chroot unshare pid namespace when possible.
This adds a helper 'unshare-pidns' that invokes unshare --pid --fork
if it is available.  It will not fail when it is not available though
as it is not expected to be present on trusty and precise.

The reason for this is so any unexpected children of the chroot process
will be killed allowing unmount to succeed.
See LP: #1645680 for example of such issues.

bzr-revno: 360.3.1