ubuntu_info: remove the path that did not require distro-info
At one point we had a functional path that did not require installation
of python-distro-info.  At some point that path broke because we were
exporting data in REL2VER was not accounted for in the __RELEASE_DATA.
Specifically missing was 'supported'.

The primary path that was enabled by the hard coded data was testing
with a new distro info before the package was sru'd.  To account for
that I've put a comment in the code showing how one can be added.

As a drive-by, this also adds the 'release_date' field to the data
exported.

bzr-revno: 369.1.1