bin/kpack-from-image: install rooturl if not already installed.
This modifies the build of the initramfs so that they include
cloud-initramfs-rooturl.  It will fail if that package is not available.

cloud-initramfs-rooturl is marked for SRU to current releases under
bug 1688606.

The code will install the package in precise if it is available but
not fail if not available.  On other releases that is fatal.

LP: #1688606
bzr-revno: 371.1.1