fix call to centos_apply_network_config to only pass in 'network' config.
This was missed as a method in curtin changed its signature to be
more clean. centos_apply_network_config never used anything but the
'network' portion of the config so it made no sense to reference
the whole config.

The maas images portion was just never changed.  This fixes that.

bzr-revno: 378.1.1