maas-cloudimg2eph2: repack squash if additional archives are used.
If --proposed or M2E_ADD_REPOS or M2E_ADD_PACKAGES are set, then that
means we should re-pack squashfs.

This check is not perfect at all, as other things might cause that, but
it is required to get a new squashfs created for the path already in
existance.

bzr-revno: 381.1.1