remove a warning that is no longer relevant.
If you ran this on precise, you'd get a warning.
The warning was also shown on bionic due to the way it compared.

Since it was only intended to show on releases older than trusty
and there are no supported releases older than trusty... it should
be safe to remove.

bzr-revno: 384