Add zfs module and dependencies to generated initramfs.
In order for curtin to use zfs, it has to have access to the modules.
Add 'zfs' and its dependencies to the initramfs.

Note that the user-space tools (zfsutils-linux) are not currently
a dependency of ubuntu-server and are thus not in the squashfs images.
That means that for curtin to use zfs it has to install the package.

LP: #1737592
bzr-revno: 385.1.1