stop grub scripts from running when updating an image.
An upgrade that brought a new kernel would cause grub's' kernel
hook scripts to run. Those would fail to run and cause failure
(at least in trusty).  This is similar to code that is currently
in the kpack-from-image, but handles putting things back more elegantly.
(kpack-from-image doesn't have to put things back).

I found this when running with '--proposed'.

bzr-revno: 386.1.1