Always create squashfs image if configured.
For precise in the v3 stream there was no 'squashfs' image created.
This was enabled only if input to the build was '.tar.gz'  or 'squashfs'.

As we need a squashfs image for installing precise with maas 2.3+,
we modify the code here to build that image.

Note that this squashfs image is known to not boot with rooturl
as the other squashfs images produced here.

LP: #1739761
bzr-revno: 387.2.1