Do not install cloud-initramfs-dyn-netconf in releases newer than xenial.
The cloud-initramfs-dyn-netconf was being added to the maas image.
It was only absolutely necessary in <xenial, but is left as present
in xenial as images still use the /etc/network/interfaces link.
See change="symlink ENI for cloud-initramfs-dyn-netconf".

It is not needed in or used without ifupdown, and ifupdown is not present
in artful or bionic.

bzr-revno: 391