Re-add cloud-initramfs-dyn-netconf.  Needed by maas for BOOTIF in ip=.
The previous commit to trunk dropped cloud-initramfs-dyn-netconf from
the maas image in releases newer than xenial.

That caused errors in MAAS. cloud-initramfs-dyn-netconf does 2 things:
 a.) writes /run/network/dynamic-interfaces
     this is a ENI style file with data retrived from ipconfig
 b.) convert 'BOOTIF' string in the value of IP
   Example:
     ip=::::foobar:BOOTIF BOOTIF=01-00-22-68-10-c1-e6
   will be effectively rendered as:
     ip=::::foobar:eth0

'b' is still required by maas.

Note: in maas v3 images, the squashfs image is re-used pristine
from cloud-images.  Because of this, the installation into the image
affects initramfs creation, but does not affect the image.  That is
correct in that this package is not needed in the image but only
in the initramfs.

LP: #1749019
bzr-revno: 392