Install CentOS tools needed to allow custom storage deployment.
Custom storage deployment requires certain tools to be installed in the target
system. If these tools are not installed Curtin will attempt to install them
using the target's package manager. This will not work in an isolated
environment so make sure they are installed when CentOS is being built.

bzr-revno: 405