Revert Open Firmware PPC64EL bootloader to Xenial.
The Bionic bootloader for Open Firmware PPC64EL is requesting a GRUB
configuration file for the wrong MAC address(LP:1785859). Until this
bug is fixed revert to the Xenial bootloader.

To allow the bootloader to be reverted lp:maas-images will now only look to
see if the bootloader discovered in the archive differs from the latest
bootloader in the stream. Older bootloaders in the stream are ignored. This
allows the release for each bootloader to be reverted to an older version
or upgraded to a newer version without changing anything other than
bootloaders.yaml.

bzr-revno: 405.1.1