Update CentOS 7 curthooks to use internal Curtin storage support.
Curtin now supports deploying custom storage on CentOS 7+. Hooks are no longer
required for CentOS but they are still included for backwards compatibility. If
CentOS is deployed with MAAS 2.5+ and a supported version of Curtin the hooks
will simply call back into Curtin to configure storage. If an older version of
MAAS or Curtin is used the old hooks will still run.

bzr-revno: 407.1.1