Cleanup: be more specific about what should be created.
This puts more information in the config files on what output is expected.
Previously we were relying on the input type format in some cases to
determine what output to create.

The desire to do this was really to safely add '--no-gzip'
to the maas-cloudimg2eph2 if we were going to throw away the root-image.gz.
The gzip of root-image.gz would typically take > 2 minutes and then
the output would be promptly thrown away.

Now, in the configs we have:
 meph-v2.yaml:
   squashfs: False
   root_image: True
   enable_di: True

 meph-v3.yaml:
   squashfs: True
   root_image: False
   enable_di: False

There is one inconsitency there that we've addressed with an 'if release'
section in the code. meph-v3.yaml will build a root image for trusty.

Also added here is information about 'known_callers' and the start of
a shell test suite. At very least it allowed me to manually verify the
output of a before and after run.

Ideally we'd have tests that assert files created and contents look good.