Add 2018-09-17 archive signing key to local keyring.gpg
maas images needs to validate signatures from trusty and later.
For that reason, we can't just use and assume that package provided
 /usr/share/keyrings/ubuntu-archive-keyring.gpg
as current versions (disco) will not include the signing key for trusty
(2012-05-11), and older versions do not include the most recent
signing key (2018-09-17).

Because we want to be able to verify data of archive content back
to trusty, we just maintain our own keyring here.