fix: use InRelease when building bootloaders
Before, we downloaded Release and its signature Release.gpg separately.
This can cause failures when the archive is served by multiple servers
and those servers are not completely in sync, or if the archive is updated
while we download the files.

InRelease is the Release file with an inline signature, which removes the
race condition.